create url.[ch], and move the functions
