+++ CMakeLists.txt
@@ -46,6 +46,7 @@ SET(LIB_SOURCE_FILES
     lib/common/string.c
     lib/common/time.c
     lib/common/timeout.c
+    lib/common/url.c
 
     lib/core/config.c
     lib/core/configurator.c
@@ -81,6 +82,7 @@ SET(UNIT_TEST_SOURCE_FILES
     t/00unit/lib/common/serverutil.c
     t/00unit/lib/common/string.c
     t/00unit/lib/common/time.c
+    t/00unit/lib/common/url.c
     t/00unit/lib/core/headers.c
     t/00unit/lib/handler/file.c
     t/00unit/lib/handler/mimemap.c
@@ -91,6 +93,7 @@ LIST(REMOVE_ITEM UNIT_TEST_SOURCE_FILES
     lib/common/serverutil.c
     lib/common/string.c
     lib/common/time.c
+    lib/common/url.c
     lib/core/headers.c
     lib/handler/file.c
     lib/handler/mimemap.c
