+++ include/h2o.h
@@ -43,6 +43,7 @@ extern "C" {
 #include "h2o/string_.h"
 #include "h2o/time_.h"
 #include "h2o/timeout.h"
+#include "h2o/url.h"
 #include "h2o/version.h"
 
 #ifndef H2O_MAX_HEADERS
