+++ include/h2o/string_.h
@@ -36,14 +36,6 @@ extern "C" {
 
 #define H2O_STRLIT(s) (s), sizeof(s) - 1
 
-typedef struct st_h2o_parse_url_t {
-    h2o_iovec_t scheme;
-    h2o_iovec_t authority; /* i.e. host:port */
-    h2o_iovec_t host;
-    h2o_iovec_t path;
-    uint16_t port;
-} h2o_parse_url_t;
-
 /**
  * duplicates given string
  * @param pool memory pool (or NULL to use malloc)
@@ -96,18 +88,6 @@ const char *h2o_next_token(h2o_iovec_t *iter, int separator, size_t *element_len
  * tests if string needle exists within a separator-separated string (for handling "#rule" of RFC 2616)
  */
 int h2o_contains_token(const char *haysack, size_t haysack_len, const char *needle, size_t needle_len, int separator);
-/**
- * removes "..", ".", decodes %xx from a path representation
- * @param pool memory pool to be used in case the path contained references to directories
- * @param path source path
- * @param len source length
- * @return buffer pointing to source, or buffer pointing to an allocated chunk with normalized representation of the given path
- */
-h2o_iovec_t h2o_normalize_path(h2o_mem_pool_t *pool, const char *path, size_t len);
-/**
- * parses absolute URL (either http or https)
- */
-int h2o_parse_url(const char *url, size_t url_len, h2o_parse_url_t *result);
 /**
  * HTML-escapes a string
  * @param pool memory pool
